Fix & update migrations

+++ db/migrate/20200610191908_drop_consent_manage_attribute_from_policy_groups_and_policy_group_templates.rb
@@ -1,6 +1,11 @@
 class DropConsentManageAttributeFromPolicyGroupsAndPolicyGroupTemplates < ActiveRecord::Migration[5.2]
   def change
-    remove_column :policy_groups, :onboarding_consent_manage, :boolean
-    remove_column :policy_group_templates, :onboarding_consent_manage, :boolean
+    if column_exists? :policy_groups, :onboarding_consent_manage
+      remove_column :policy_groups, :onboarding_consent_manage, :boolean
+    end
+
+    if column_exists? :policy_group_templates, :onboarding_consent_manage
+      remove_column :policy_group_templates, :onboarding_consent_manage, :boolean
+    end
   end
 end
